Default message

+++ transformers/modeling_bert.py
@@ -206,7 +206,7 @@ class BertSelfAttention(nn.Module):
 
         if self.add_cnn:
             print("ADDING CNN")
-            padding = (config.cnn_filter_width - 1) / 2
+            padding = int((config.cnn_filter_width - 1) / 2)
             self.convolutions = []
             for i in range(self.num_attention_heads):
                 self.convolutions.append(nn.Conv1d(config.max_seq_length,
